+++ src/IECoreDelight/Renderer.cpp
@@ -266,7 +266,18 @@ class DelightOutput : public IECore::RefCounted
 			const char *typePtr = output->getType().c_str();
 			const char *namePtr = output->getName().c_str();
 
-			ParameterList driverParams( output->parameters() );
+			ParameterList driverParams;
+			for( const auto &[parameterName, parameterValue] : output->parameters() )
+			{
+				// We can't pass `filter` to the driver, because although it's not
+				// documented as an attribute (and it _is_ documented that additional
+				// arbitrary attributes are allowed), 3Delight complains.
+				if( parameterName != "filter" )
+				{
+					driverParams.add( parameterName.c_str(), parameterValue.get() );
+				}
+			}
+
 			driverParams.add( { "drivername", &typePtr, NSITypeString, 0, 1, 0 } );
 			driverParams.add( { "imagefilename", &namePtr, NSITypeString, 0, 1, 0 } );
 
@@ -277,6 +288,7 @@ class DelightOutput : public IECore::RefCounted
 			string variableName;
 			string variableSource;
 			string layerType;
+			string scalarFormat;
 			string layerName;
 			int withAlpha = 0;
 
@@ -303,6 +315,11 @@ class DelightOutput : public IECore::RefCounted
 				if( tokens[0] == "float" )
 				{
 					layerType = "scalar";
+				}
+				else if( tokens[0] == "uint" )
+				{
+					layerType = "scalar";
+					scalarFormat = "uint32";
 				}
 				else if( tokens[0] == "point" )
 				{
@@ -333,6 +350,27 @@ class DelightOutput : public IECore::RefCounted
 				boost::replace_all( layerName, ".", "_" );
 			}
 
+			// Special cases to match the "standard" expected by OutputBuffer,
+			// which is necessary for Gaffer viewport rendering.
+
+			if( variableName == "Z" )
+			{
+				variableName = "z";
+				variableSource = "builtin";
+			}
+
+			if( variableName == "id" )
+			{
+				variableName = "cortexId";
+				variableSource = "attribute";
+				/// \todo We really want to use something like "uint32" here (as
+				/// provided by the code above), but that maps the `0.0 - 1.0`
+				/// range into the integer range, whereas we want a direct
+				/// mapping. So we render as float and deal with it in
+				/// Display.cpp.
+				scalarFormat = "float";
+			}
+
 			layerName = parameter<string>( output->parameters(), "layerName", layerName );
 
 			ParameterList layerParams;
@@ -343,10 +381,21 @@ class DelightOutput : public IECore::RefCounted
 			layerParams.add( "layername", layerName );
 			layerParams.add( { "withalpha", &withAlpha, NSITypeInteger, 0, 1, 0 } );
 
-			const string scalarFormat = this->scalarFormat( output );
-			const string colorProfile = scalarFormat == "float" ? "linear" : "sRGB";
+			string colorProfile = "linear";
+			if( scalarFormat.empty() )
+			{
+				scalarFormat = this->scalarFormat( output );
+				colorProfile = scalarFormat == "float" ? "linear" : "sRGB";
+			}
 			layerParams.add( "scalarformat", scalarFormat );
 			layerParams.add( "colorprofile", colorProfile );
+
+			string filter = parameter<string>( output->parameters(), "filter", "blackman-harris" );
+			if( filter == "closest" )
+			{
+				filter = "zmin";
+			}
+			layerParams.add( "filter", filter );
 
 			m_layerHandle = DelightHandle( context, "outputLayer:" + name, ownership, "outputlayer", layerParams );